feat: working splash
