+++ app/src/state/swap/hooks.ts
@@ -1,11 +1,6 @@
 import { useCallback, useEffect, useState } from 'react'
-import { parseUnits } from '@ethersproject/units'
 import { useDispatch, useSelector } from 'react-redux'
-import { Currency, CurrencyAmount, Token, TokenAmount } from '@uniswap/sdk'
-import { useWeb3React } from '@web3-react/core'
 import { AppDispatch, AppState } from '../index'
-import { BigNumber } from 'ethers'
-
 import { typeInput, setLoading } from './actions'
 
 export function useSwap(): AppState['swap'] {
@@ -39,14 +34,4 @@ export function useSwapActionHandlers(): {
     [dispatch]
   )
   return { onUserInput }
-}
-
-// try to parse a user entered amount for a given token
-export function tryParseAmount(value?: string): BigNumber | undefined {
-  if (!value || value === '0' || value === '.' || value.indexOf('0') === 0)
-    return parseUnits('0', 18)
-  const typedValueParsed = parseUnits(value, 18).toString()
-  if (typedValueParsed !== '0') {
-    return BigNumber.from(typedValueParsed)
-  }
 }
