+++ app/src/utils/index.ts
@@ -1,6 +1,7 @@
 import BigNumber from 'bignumber.js'
 
 export { default as formatAddress } from './formatAddress'
+export { tryParseAmount } from './tryParseAmount'
 
 export const bnToDec = (bn: BigNumber, decimals = 18) => {
   return bn.dividedBy(new BigNumber(10).pow(decimals)).toNumber()
